Shape

+++ src/com/ust/LP2/c4/AbstractionAndInterfaces/Shape.java
@@ -1,6 +1,6 @@
 package com.ust.LP2.c4.AbstractionAndInterfaces;
 
 public interface Shape {
-    int calculateArea();
+    double calculateArea();
     void displayShapeInfo();
 }
